requirements.txt
numpy
opencv-python
pyrealsense2
roboflow
xarm-python-sdk